searchview custom, MapListview

+++ dabang/app/src/main/java/com/softsquared/template/src/main/Map/MyItem.java
@@ -11,7 +11,7 @@ public class MyItem implements ClusterItem {
     private final LatLng mPosition;
 
 
-    public MyItem(double lat, double lng) {
+    public MyItem(float lat, float lng) {
        mPosition = new LatLng(lat,lng);
     }
